Bump Build Spec to 0.2
The latest version of the Build Spec maintains state between commands
rather than running all commands in a separate install of the shell.
This allows us to set the tag as an environment variable rather than
using temporary files.

Note we're still generating a JSON file at the end of the build to pass
along build details to CloudFormation in the next step in the pipeline.