Support additional regions
CodePipeline requires S3 buckets used as source artifacts to exist in
the same Region as the pipeline. Previously the bucket existed in
us-east-1 so that's the only region the reference architecture was
deployable to. This change enhances the deployment routine to deploy to
buckets in all eligible Regions where all required services exist.