Add AWS Fargate support
Support the Fargate launch type via a parameter in the CloudFormation
template for launch type. If Amazon EC2 is specified, the template will
continue to create an Auto Scaling group of EC2 instance. If AWS Fargate
is specified, this will skipped and the service will specify FARGATE as
the launch type.

- Add service and update task definition for Fargate
- Add a log group to assist in debugging for Fargate tasks
- Convert vestigial embedded JSON into YAML
- Update reference architecture diagrams
- As we're no longer using Amazon S3 in CodePipeline, consolidate into a
  single bucket for templates
- Update the README and minor cleanup